Move the implementations of encoding_overhead() etc back to the Python file
I commented to contributor robamu that it would be good to put the functions
in _cobs_py.py etc, instead of __init__.py, and then add an equivalent C API
implementation in the .c files. However --

* robamu doesn't have experience with the Python C API.
* I haven't used the Python C API for years.
* These functions aren't as calculation-intensive as the encode/decode
  functions are, and aren't expected to be a hotspot in typical usage.

So for now, I'll just move these functions back into __init__.py.

This pertains to pull request #3.